Fix compiler errors around state usage
`STATE.borrow(cs)` returns `&State`, which doesn't have a `borrow_mut`
method. `RefCell` has a `borrow_mut`, and we do need a `RefCell` here,
but instead of adding it right now, I wanted to fix the compiler errors
in the way I did here, as that exposes a different problem.

The state variables are of type `&State`, and we're assigning another
`&State` to them. While this compiles, it never actually changes the
state that's stored in the mutex! With the changes in this commit, this
is plainly visible, as Rust warns us about the values assigned to those
variables never being read.