Move transfer into `State` enum
This takes care of one of the remaining build errors, but it's also the
better approach overall. We're already keeping track of the current
state, so why not make use of that and use the same mechanism to track
our state-specific data too. This way, there's no need to fiddle around
with options.